Add global color table parsing
The global color table is an optional part of a gif file.
This means we needed to add support for an optional token.

Additionally, it can have a variable size.
This means we needed to add support for variable sized tokens.

Lastly, to determine whether to parse a global color table or not
we needed to keep state while parsing. We were already doing this
anyway with the CurrentOffset. But it wasn't a formal concept.
There is now a parsing state that is clear and separate.

ISSUE #10